Modell fertig

+++ src/main/java/htl/steyr/cartclient/dto/request/ProductRequestDto.java
@@ -1,4 +1,4 @@
 package htl.steyr.cartclient.dto.request;
 
-public record ProductRequestDto(Long id, Long amount) {
+public record ProductRequestDto(Long id, int amount) {
 }
